requirements.txt
openai-whisper
langchain
pydub
openai
ffmpeg-python
pyicloud
ics
email
ssl
